+++ Pandora/Paradigm/Primary/Linear.hs
@@ -1,0 +1,3 @@
+module Pandora.Paradigm.Linear (module Exports) where
+
+import Pandora.Paradigm.Linear.Vector as Exports