+++ Pandora/Pattern/Linear.hs
@@ -1,3 +1,0 @@
-module Pandora.Pattern.Linear (module Exports) where
-
-import Pandora.Pattern.Linear.Vector as Exports