+++ CHANGELOG.md
@@ -410,3 +410,4 @@
 * Change precedence for `.` infix operator
 * Change precedence for `*`, `+`, `-` infix operators
 * Change precedence for `#` infix operator
+* Move `Linear` module to `Paradigm.Primary`